# docs/industrial/openapi.json
{
  "components": {
    "schemas": {
...
      "DeletionView": {
        "properties": {
          "deletion_id": {
            "format": "uuid",
            "title": "Deletion Id",
            "type": "string"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "restore_before": {
            "format": "date-time",
            "title": "Restore Before",
            "type": "string"
          },
          "state": {
            "enum": [
              "deleted",
              "purged"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "deletion_id",
          "state",
          "restore_before"
        ],
        "title": "DeletionView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "in": "cookie",
        "name": "__Host-tracedesk-session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "TraceDesk Team API",
    "version": "0.3.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/ask": {
      "post": {
        "operationId": "upgrade_api_ask_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Upgrade"
      }
    },
    "/api/demo": {
      "post": {
        "operationId": "upgrade_api_demo_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Upgrade"
      }
    },
    "/api/documents": {
      "post": {
        "operationId": "upgrade_api_documents_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Upgrade"
      }
    },
    "/api/documents/{doc_id}": {
      "delete": {
        "operationId": "upgrade_api_documents__doc_id__delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Upgrade"
      }
    },
    "/api/evaluate": {
      "post": {
        "operationId": "upgrade_api_evaluate_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Upgrade"
      }
    },
    "/api/index": {
      "post": {
        "operationId": "upgrade_api_index_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Upgrade"
      }
    },
    "/api/library": {
      "get": {
        "operationId": "library_api_library_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Library Api Library Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Library"
      }
    },
    "/api/sources/{doc_id}": {
      "get": {
        "operationId": "source_api_sources__doc_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Source Api Sources  Doc Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Source"
      }
    },
    "/api/traces/{trace_id}": {
      "get": {
        "operationId": "trace_api_traces__trace_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Trace Api Traces  Trace Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Trace"
      }
    },
...
    "/api/v1/documents/{document_id}": {
      "delete": {
        "operationId": "delete_document_api_v1_documents__document_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Delete Document"
      }
    },
...
  }
}